feat: move selectParamFromFilter into state/selectors (#3150)
* feat: move selectParamFromFilter into state/selectors

* feat: move selectXYZParamFromFilter away
