+++ src/state/slices/txHistorySlice/selectors.ts
@@ -5,12 +5,10 @@ import createCachedSelector from 're-reselect'
 import { createSelector } from 'reselect'
 import type { ReduxState } from 'state/reducer'
 import { createDeepEqualOutputSelector } from 'state/selector-utils'
+import { selectChainIdParamFromFilter } from 'state/selectors'
 
 import type { AccountMetadata } from '../portfolioSlice/portfolioSliceCommon'
-import {
-  selectChainIdParamFromFilter,
-  selectPortfolioAccountMetadata,
-} from '../portfolioSlice/selectors'
+import { selectPortfolioAccountMetadata } from '../portfolioSlice/selectors'
 import type { Tx, TxId } from './txHistorySlice'
 
 export const selectTxs = createDeepEqualOutputSelector(
